Fixes for OpenZeppelin Security Audit H01:
When the administrator calls reward or rewardAndOpenNextDraw, the secret and salt that will determine the lottery winner is revealed.

However, the selected address will depend on the distribution of committed draws, which can be influenced by sending pool tokens to another address, burning pool tokens, or withdrawing an address’ entire balance. It can also be influenced by sending SAI pool tokens to the DAI pool contract.

Mitigation:

- Token movements are locked prior to awarding the prize.
- Lock has a timeout, so that the pool can't be locked forever
- Lock has a cooldown, to prevent nasty admins from continuously locking
  the contract